Moved monitoring tests from tests_monitoring.py to tests.py
